remove is.na(X) check for disease cols
For each column in `all_disease_cols`, for each row of the table, replace the cell with a TRUE/FALSE representing whether or not the value at that exact cell was NA. This works for the old code, because it was only "1" and "NA", but does not work here. For example, this would convert the following column:

TRUE
TRUE
NA
FALSE


TRUE
TRUE
FALSE
TRUE

in the above way. THIS IS VERY BAD AND HORRIBLE